+++ packages/dataviews/src/dataviews-layouts/activity/activity-group.tsx
@@ -10,19 +10,19 @@ import { createInterpolateElement } from '@wordpress/element';
  */
 import type { NormalizedField } from '../../types';
 
-interface TimelineGroupProps< Item > {
+interface ActivityGroupProps< Item > {
 	groupName: string;
 	groupData: Item[];
 	groupField: NormalizedField< Item >;
 	children: React.ReactNode;
 }
 
-export default function TimelineGroup< Item >( {
+export default function ActivityGroup< Item >( {
 	groupName,
 	groupData,
 	groupField,
 	children,
-}: TimelineGroupProps< Item > ) {
+}: ActivityGroupProps< Item > ) {
 	// Determine if we should show the field label
 	const groupHeader = createInterpolateElement(
 		// translators: %s: The label of the field e.g. "Status".
@@ -41,9 +41,9 @@ export default function TimelineGroup< Item >( {
 		<VStack
 			key={ groupName }
 			spacing={ 0 }
-			className="dataviews-view-timeline__group"
+			className="dataviews-view-activity__group"
 		>
-			<h3 className="dataviews-view-timeline__group-header">
+			<h3 className="dataviews-view-activity__group-header">
 				{ groupHeader }
 			</h3>
 			{ children }